refactor: share types schemas and utilities
