+++ src/components/Lesson/LessonDownloads/LessonDownloads.page.tsx
@@ -13,7 +13,7 @@ import useResourceFormSubmit from "@/components/DownloadAndShareComponents/hooks
 import {
   ResourceFormProps,
   DownloadResourceType,
-} from "@/components/DownloadAndShareComponents/downloadsAndShare.types";
+} from "@/components/DownloadAndShareComponents/downloadAndShare.types";
 import Breadcrumbs from "@/components/Breadcrumbs";
 import DownloadCardGroup from "@/components/DownloadAndShareComponents/DownloadCardGroup/DownloadCardGroup";
 import FieldError from "@/components/FormFields/FieldError";
